# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PulseBeat AI Health Application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupRequest {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserResponse {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: UserResponse;
  message?: string;
}

export type ScanResultMood =
  (typeof ScanResultMood)[keyof typeof ScanResultMood];

export const ScanResultMood = {
  Relaxed: "Relaxed",
  Normal: "Normal",
  Excited: "Excited",
  Stressed: "Stressed",
} as const;

export interface ScanResult {
  /** Heart rate in BPM */
  heartRate: number;
  /** Respiratory rate in breaths per minute */
  respiratoryRate: number;
  /** Heart rate variability SDNN */
  hrvSdnn?: number;
  mood: ScanResultMood;
  /** PPG waveform data points */
  ppgWaveform?: number[];
  /** Analysis confidence score 0-1 */
  confidence: number;
  /** Saved scan record ID */
  scanId?: number;
}

export type ScanRecordMood =
  (typeof ScanRecordMood)[keyof typeof ScanRecordMood];

export const ScanRecordMood = {
  Relaxed: "Relaxed",
  Normal: "Normal",
  Excited: "Excited",
  Stressed: "Stressed",
} as const;

export interface ScanRecord {
  id: number;
  heartRate: number;
  respiratoryRate: number;
  hrvSdnn?: number;
  mood: ScanRecordMood;
  confidence?: number;
  createdAt: string;
}

export interface MusicTrack {
  videoId: string;
  title: string;
  channelTitle: string;
  thumbnail: string;
  duration?: string;
}

export interface MusicRecommendationsResponse {
  mood: string;
  genre: string;
  tracks: MusicTrack[];
}

export type DashboardStatsMoodDistribution = { [key: string]: number };

export type DashboardStatsHeartRateTrendItem = {
  date: string;
  heartRate: number;
};

export interface DashboardStats {
  totalScans: number;
  averageHeartRate: number;
  averageRespiratoryRate: number;
  moodDistribution: DashboardStatsMoodDistribution;
  recentScans: ScanRecord[];
  heartRateTrend: DashboardStatsHeartRateTrendItem[];
}

export type AnalyzeScanBody = {
  /** Recorded facial video file */
  video: Blob;
};

export type GetMusicRecommendationsParams = {
  mood: GetMusicRecommendationsMood;
};

export type GetMusicRecommendationsMood =
  (typeof GetMusicRecommendationsMood)[keyof typeof GetMusicRecommendationsMood];

export const GetMusicRecommendationsMood = {
  Relaxed: "Relaxed",
  Normal: "Normal",
  Excited: "Excited",
  Stressed: "Stressed",
} as const;