# lib/api-zod/src/generated/types/musicTrack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PulseBeat AI Health Application API
 * OpenAPI spec version: 0.1.0
 */

export interface MusicTrack {
  videoId: string;
  title: string;
  channelTitle: string;
  thumbnail: string;
  duration?: string;
}