# lib/api-zod/src/generated/types/scanRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PulseBeat AI Health Application API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanRecordMood } from "./scanRecordMood";

export interface ScanRecord {
  id: number;
  heartRate: number;
  respiratoryRate: number;
  hrvSdnn?: number;
  mood: ScanRecordMood;
  confidence?: number;
  createdAt: Date;
}