# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PulseBeat AI Health Application API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanResultMood } from "./scanResultMood";

export interface ScanResult {
  /** Heart rate in BPM */
  heartRate: number;
  /** Respiratory rate in breaths per minute */
  respiratoryRate: number;
  /** Heart rate variability SDNN */
  hrvSdnn?: number;
  mood: ScanResultMood;
  /** PPG waveform data points */
  ppgWaveform?: number[];
  /** Analysis confidence score 0-1 */
  confidence: number;
  /** Saved scan record ID */
  scanId?: number;
}